+++ zincio-cegla/src/cgi.rs
@@ -2,19 +2,19 @@ use std::{path::PathBuf, process::Stdio};
 
 use cegla_cgi::CgiEnvironment;
 
-/// `vibeio`-based runtime for `cegla-cgi`
-pub struct VibeioCgiRuntime;
+/// `zincio`-based runtime for `cegla-cgi`
+pub struct ZincioCgiRuntime;
 
-/// `vibeio`-based child process for `cegla-cgi`
-pub struct VibeioCgiChild {
-  inner: vibeio::process::Child,
+/// `zincio`-based child process for `cegla-cgi`
+pub struct ZincioCgiChild {
+  inner: zincio::process::Child,
 }
 
-impl cegla_cgi::client::Runtime for VibeioCgiRuntime {
-  type Child = VibeioCgiChild;
+impl cegla_cgi::client::Runtime for ZincioCgiRuntime {
+  type Child = ZincioCgiChild;
 
   fn spawn(&self, future: impl std::future::Future + 'static) {
-    vibeio::spawn(async move {
+    zincio::spawn(async move {
       future.await;
     });
   }
@@ -26,7 +26,7 @@ impl cegla_cgi::client::Runtime for VibeioCgiRuntime {
     env: CgiEnvironment,
     cwd: Option<PathBuf>,
   ) -> Result<Self::Child, std::io::Error> {
-    let mut command = vibeio::process::Command::new(cmd);
+    let mut command = zincio::process::Command::new(cmd);
     command
       .stdin(Stdio::piped())
       .stdout(Stdio::piped())
@@ -36,27 +36,27 @@ impl cegla_cgi::client::Runtime for VibeioCgiRuntime {
     if let Some(cwd) = cwd {
       command.current_dir(cwd);
     }
-    Ok(VibeioCgiChild {
+    Ok(ZincioCgiChild {
       inner: command.spawn()?,
     })
   }
 }
 
-impl cegla_cgi::client::Child for VibeioCgiChild {
-  type Stdin = vibeio::util::AsyncWrap<vibeio::process::ChildStdin>;
-  type Stdout = vibeio::util::AsyncWrap<vibeio::process::ChildStdout>;
-  type Stderr = vibeio::util::AsyncWrap<vibeio::process::ChildStderr>;
+impl cegla_cgi::client::Child for ZincioCgiChild {
+  type Stdin = zincio::util::AsyncWrap<zincio::process::ChildStdin>;
+  type Stdout = zincio::util::AsyncWrap<zincio::process::ChildStdout>;
+  type Stderr = zincio::util::AsyncWrap<zincio::process::ChildStderr>;
 
   fn stdin(&mut self) -> Option<Self::Stdin> {
-    self.inner.stdin.take().map(vibeio::util::AsyncWrap::new)
+    self.inner.stdin.take().map(zincio::util::AsyncWrap::new)
   }
 
   fn stdout(&mut self) -> Option<Self::Stdout> {
-    self.inner.stdout.take().map(vibeio::util::AsyncWrap::new)
+    self.inner.stdout.take().map(zincio::util::AsyncWrap::new)
   }
 
   fn stderr(&mut self) -> Option<Self::Stderr> {
-    self.inner.stderr.take().map(vibeio::util::AsyncWrap::new)
+    self.inner.stderr.take().map(zincio::util::AsyncWrap::new)
   }
 
   fn try_status(&mut self) -> std::io::Result<Option<std::process::ExitStatus>> {