+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## `cegla` UNRELEASED
+
+- Replaced `vibeio-cegla` with `zincio-cegla`, following the rename of `vibeio` to `zincio`.
+
 ## `cegla` 0.2.3 - April 25, 2026
 
 - Fixed EOF not being sent for error stream when the client receives FastCGI END_REQUEST record.